Refactor: move TestApp class out of conftest.py
