+++ repondeur/tests/conftest.py
@@ -1,5 +1,4 @@
 import os
-from contextlib import contextmanager
 from pathlib import Path
 
 import pytest
@@ -18,44 +17,10 @@ from fixtures.plfss2019 import *  # noqa: F401,F403
 from fixtures.scraping import *  # noqa: F401,F403
 from fixtures.shared_tables import *  # noqa: F401,F403
 from fixtures.users import *  # noqa: F401,F403
-from testapp import TestApp as BaseTestApp
+from testapp import TestApp
 from visam.fixtures.conseils import *  # noqa: F401,F403
 
 HERE = Path(__file__)
-
-
-class TestApp(BaseTestApp):
-    def get(self, *args, **kwargs):
-        with self.auto_login(kwargs):
-            return super().get(*args, **kwargs)
-
-    def post(self, *args, **kwargs):
-        with self.auto_login(kwargs):
-            return super().post(*args, **kwargs)
-
-    def post_json(self, *args, **kwargs):
-        with self.auto_login(kwargs):
-            return super().post_json(*args, **kwargs)
-
-    @contextmanager
-    def auto_login(self, kwargs):
-        from zam_repondeur.models import User
-
-        user = kwargs.pop("user", None)
-        if user is not None:
-            assert isinstance(user, User)
-            self.user_login(email=user.email, headers=kwargs.get("headers"))
-
-        yield
-
-    def user_login(self, email, headers=None):
-        from zam_repondeur.auth import generate_auth_token
-        from zam_repondeur.services.users import repository
-
-        token = generate_auth_token()
-        repository.set_auth_token(email, token)
-        resp = self.get("/authentification", params={"token": token}, headers=headers)
-        assert resp.status_code == 302
 
 
 @pytest.fixture(scope="session")
